Use file extension based on the file format
